+++ R/gzipr.R
@@ -25,6 +25,8 @@
 #'   `y` is used as labels for `x`, otherwise, if `x` is named
 #'   (`rownames` for `data.frame`s, `names` for `character`s) `y` is set
 #'   to the names of `x`. If `x` is not named, `y` must be provided.
+#' @param test (default = `FALSE`) whether the input data is for test
+#'   (`TRUE`, i.e., without labels) or for train (`FALSE`)
 #'
 #' @return a `gzipr` model object
 #' @export
@@ -50,57 +52,78 @@
 #'
 #' # accuracy
 #' mean(result == test_y)
-gzipr <- function(x, y = NULL) {
+gzipr <- function(x, y = NULL, test = FALSE) {
   UseMethod("gzipr")
 }
 
 #' @describeIn gzipr method to train on `character` vectors as input
 #'   data
 #' @export
-gzipr.character <- function(x, y = NULL) {
-  as.list(x) |>
-    purrr::set_names(y %||% x) |>
-    structure(class = "gzipr")
+gzipr.character <- function(x, y = NULL, test = FALSE) {
+  x <- as.list(x)
+
+  y <- y %||% names(x)
+
+  if (is.null(y)) {
+    usethis::ui_stop("y must be provided if x has no names")
+  }
+
+  if (!test) {
+    x <- purrr::set_names(x, y)
+  }
+
+  structure(x, class = "gzipr")
 }
 
 #' @describeIn gzipr method to train on `data.frame`s as input data
 #' @export
-gzipr.data.frame <- function(x, y = NULL) {
+gzipr.data.frame <- function(x, y = NULL, test = FALSE) {
   nrow_x <- nrow(x)
   y <- y %||% rownames(x)
-  if (
-    is.null(y) ||
+
+  x <- seq_len(nrow_x) |>
+    purrr::map(\(id_row) x[id_row, , drop = FALSE])
+
+  if (!test) {
+    if (
+      is.null(y) ||
       isTRUE(all.equal(y, as.character(seq_len(nrow_x))))
-  ) {
-    usethis::ui_stop(paste(
-      "y must be provided if x has no or standard rownames",
-      "(i.e., seq_len(nrow(x)))."
-    ))
+    ) {
+      usethis::ui_stop(paste(
+        "y must be provided if x has no or standard rownames",
+        "(i.e., seq_len(nrow(x)))."
+      ))
+    }
+    x <- purrr::set_names(x, y)
   }
-  seq_len(nrow_x) |>
-    purrr::map(\(id_row) x[id_row, , drop = FALSE]) |>
-    purrr::set_names(y) |>
-    structure(class = "gzipr")
+
+  structure(x, class = "gzipr")
 }
 
 #' @describeIn gzipr `r lifecycle::badge("experimental")`
 #'   method to train on `list`s as input data
 #' @export
-gzipr.list <- function(x, y = NULL) {
+gzipr.list <- function(x, y = NULL, test = FALSE) {
   stopifnot(length(unique(purrr::map(x, class))) == 1L)
-  purrr::set_names(x, y) |>
-    structure(class = "gzipr")
+
+  if (!test) {
+    x <- purrr::set_names(x, y)
+  }
+  structure(x, class = "gzipr")
 }
 
 #' @describeIn gzipr method to retrain a `gzipr` model maintaining the
 #'   same data and possibly changing the labels
 #' @export
-gzipr.gzipr <- function(x, y = NULL) {
+gzipr.gzipr <- function(x, y = NULL, test = FALSE) {
   if (!is.null(names(x))) {
     y <- y %||% names(x)
   }
-  purrr::set_names(unname(x), y) |>
-    structure(class = "gzipr")
+
+  if (!test) {
+    x <- purrr::set_names(unname(x), y)
+  }
+  structure(x, class = "gzipr")
 }
 
 #' @describeIn gzipr catch-all method for not yet implemented training
@@ -117,6 +140,6 @@ gzipr.default <- function(x, y = NULL) {
 predict.gzipr <- function(gzipr, newdata = NULL, k = 3) {
   stopifnot(is_gzipr(gzipr))
 
-  newdata <- gzipr(newdata %||% unname(gzipr))
+  newdata <- gzipr(newdata %||% unname(gzipr), test = TRUE)
   purrr::map_chr(newdata, \(x1) gzip_knn(x1, gzipr, k = k))
 }