+++ NEWS.md
@@ -1,5 +1,7 @@
 # gzipr (development version)
 
+* Fix a bug that caused the `predict()` function to fail when the new input data to predict was without rownames.
+
 # gzipr 0.1.0
 
 * Added a `gzipr()` function to train a model. Including support for tabular (`data.frame`) and text (`character`) data. Input `list` is also supported but questioning and experimental only.